+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+## 0.0.3
+
+* Add vertical divider
+* Replace verticalPadding with padding
+
 ## 0.0.2
 
-* Improve document
+* Improve documents
 
 ## 0.0.1